svn merge -r 31316:31907 https://dart.googlecode.com/svn/branches/bleeding_edge trunk

git-svn-id: http://dart.googlecode.com/svn/trunk@31918 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/datastream.h
@@ -158,7 +158,7 @@ class WriteStream : public ValueObject {
   }
 
   uint8_t* buffer() const { return *buffer_; }
-  int bytes_written() const { return current_ - *buffer_; }
+  intptr_t bytes_written() const { return current_ - *buffer_; }
 
   void set_current(uint8_t* value) { current_ = value; }
